moved Cards component to /common,created dynamic heading for table, deleted Test component, removed unused imports
