+++ src/Constant/table.js
@@ -1,0 +1,19 @@
+export const homeTableHeading = [
+  "EMPLOYEE NAME",
+  "EMPLOYEE ID",
+  "LEVEL ACCESS",
+  "STATUS",
+  "CHECK-IN",
+  "CHECK-OUT",
+  "CEHCK INTERVAL",
+];
+
+export const singleScreenTableHeading = [
+  "HOST CAPTION",
+  "STATE",
+  "STATUS",
+  "RESPONSE",
+  "CHECK-IN TIME",
+  "CHECK-OUT TIME",
+  "CHECK INTERVAL",
+];
